requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
Flask-Migrate==4.0.5
python-dotenv==1.0.0
google-generativeai==0.3.2
google-cloud-documentai==3.7.0
#Pillow==10.1.0
requests==2.31.0
gunicorn==21.2.0
#psycopg2-binary==2.9.9
pytest==7.4.2
python-dateutil==2.8.2
cryptography==41.0.4
Werkzeug==3.0.0
click==8.1.7
Marshmallow==3.20.1
openpyxl==3.1.5
#pandas==2.1.1
sendgrid==6.10.0
